+++ tidings/models.py
@@ -64,8 +64,8 @@ class Watch(ModelBase):
         # TODO: Trace event_type back to find the Event subclass, and ask it
         # how to describe me in English.
         rest = self.content_object or self.content_type or self.object_id
-        return u'Watch %s: %s, %s' % (self.pk, self.event_type,
-                                         unicode(rest))
+        return u'id=%s, type=%s, content_object=%s' % (self.pk, self.event_type,
+                                                       unicode(rest))
 
     def activate(self):
         """Enable this watch so it actually fires.